+++ scripts/sql_s27_batch.sql
@@ -1,0 +1,88 @@
+BEGIN;
+
+-- Crusader Tripoli (id=471) - 1102-1289
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(471, 'Tyerman, C. (2006). God''s War: A New History of the Crusades. Harvard University Press / Allen Lane. ISBN 978-0674030701.', 'https://www.hup.harvard.edu/catalog.php?isbn=9780674030701', 'academic'),
+(471, 'Madden, T. F. (2005). The New Concise History of the Crusades. Rowman & Littlefield. ISBN 978-0742538221.', 'https://rowman.com/ISBN/9780742538221', 'academic'),
+(471, 'Barber, M. (2012). The Crusader States. Yale University Press. ISBN 978-0300189315.', 'https://yalebooks.yale.edu/9780300189315', 'academic'),
+(471, 'Kennedy, H. (1994). Crusader Castles. Cambridge University Press. ISBN 978-0521799133.', 'https://www.cambridge.org/9780521799133', 'academic'),
+(471, 'Riley-Smith, J. (2012). The Knights Hospitaller in the Levant, c.1070-1309. Palgrave Macmillan. ISBN 978-0230290839.', 'https://link.springer.com/book/10.1057/9781137264008', 'academic');
+
+-- Crusader Edessa (id=473) - 1098-1150
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(473, 'Ferdinandi, S. (2017). La Contea Franca di Edessa: Fondazione e Profilo Storico del Primo Principato Crociato nel Levante (1098-1150). Pontificia Università Antonianum. ISBN 978-8872571033.', 'https://www.antonianum.eu', 'academic'),
+(473, 'MacEvitt, C. (2008). The Crusades and the Christian World of the East: Rough Tolerance. University of Pennsylvania Press. ISBN 978-0812240504.', 'https://www.upenn.edu/pennpress/book/14422.html', 'academic'),
+(473, 'Tyerman, C. (2006). God''s War: A New History of the Crusades. Harvard University Press. ISBN 978-0674030701.', 'https://www.hup.harvard.edu/catalog.php?isbn=9780674030701', 'academic'),
+(473, 'Barber, M. (2012). The Crusader States. Yale University Press. ISBN 978-0300189315.', 'https://yalebooks.yale.edu/9780300189315', 'academic'),
+(473, 'Amouroux-Mourad, M. (1988). Le Comté d''Édesse 1098-1150. Geuthner. ISBN 978-2705305444.', 'https://www.editions-geuthner.com', 'academic');
+
+-- Melayu Kingdom ملايو (id=866) - 671-1347 Sumatra
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(866, 'Cœdès, G. (1968). The Indianized States of Southeast Asia. University of Hawai''i Press. ISBN 978-0824803681.', 'https://uhpress.hawaii.edu/title/the-indianized-states-of-southeast-asia/', 'academic'),
+(866, 'Munoz, P. M. (2006). Early Kingdoms of the Indonesian Archipelago and the Malay Peninsula. Editions Didier Millet. ISBN 978-9814155670.', 'https://www.edmbooks.com', 'academic'),
+(866, 'Miksic, J. N. & Goh, G. Y. (2017). Ancient Southeast Asia. Routledge. ISBN 978-0415735544.', 'https://www.routledge.com/9780415735544', 'academic'),
+(866, 'I-Tsing / Takakusu, J. (transl.) (2000). A Record of the Buddhist Religion As Practised in India and the Malay Archipelago (A.D. 671-695). Munshiram Manoharlal. ISBN 978-8120616226.', 'https://www.mrmlbooks.com', 'academic'),
+(866, 'Manguin, P.-Y. (2002). The Amorphous Nature of Coastal Polities in Insular Southeast Asia: Restricted Centres, Extended Peripheries. Moussons, 5, 73-99. DOI:10.4000/moussons.2699.', 'https://doi.org/10.4000/moussons.2699', 'academic');
+
+-- Tlaxcallan (id=643) - 1348-1525 Mexica resistance
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(643, 'Gibson, C. (1952). Tlaxcala in the Sixteenth Century. Yale University Press / Stanford University Press reprint 1967. ISBN 978-0804703697.', 'https://www.sup.org/books/title/?id=12058', 'academic'),
+(643, 'Fargher, L. F., Blanton, R. E. & Heredia Espinoza, V. Y. (2010). Egalitarian Ideology and Political Power in Prehispanic Central Mexico: The Case of Tlaxcallan. Latin American Antiquity, 21(3), 227-251. DOI:10.7183/1045-6635.21.3.227.', 'https://doi.org/10.7183/1045-6635.21.3.227', 'academic'),
+(643, 'Restall, M. & Asselbergs, F. (2007). Invading Guatemala: Spanish, Nahua, and Maya Accounts of the Conquest Wars. Pennsylvania State University Press. ISBN 978-0271027586.', 'https://www.psupress.org/books/titles/978-0-271-02758-6.html', 'academic'),
+(643, 'Muñoz Camargo, D. (1585/1998). Historia de Tlaxcala (ed. L. Reyes García). Universidad Autónoma de Tlaxcala / CIESAS. ISBN 978-9686340761.', 'https://www.uatx.mx', 'primary'),
+(643, 'Hassig, R. (1988). Aztec Warfare: Imperial Expansion and Political Control. University of Oklahoma Press. ISBN 978-0806121215.', 'https://www.oupress.com/9780806121215/aztec-warfare', 'academic');
+
+-- Nogay Horde (id=338) - 1390-1634 steppe
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(338, 'Khodarkovsky, M. (2002). Russia''s Steppe Frontier: The Making of a Colonial Empire, 1500-1800. Indiana University Press. ISBN 978-0253217707.', 'https://iupress.org/9780253217707/russias-steppe-frontier/', 'academic'),
+(338, 'Davies, B. L. (2007). Warfare, State and Society on the Black Sea Steppe, 1500-1700. Routledge. ISBN 978-0415239868.', 'https://www.routledge.com/9780415239868', 'academic'),
+(338, 'Trepavlov, V. V. (2001). История Ногайской Орды [History of the Nogai Horde]. Vostochnaya Literatura. ISBN 978-5020180888.', 'https://www.vostlit.info', 'academic'),
+(338, 'Başer, A. (2019). Conflicting Legitimacies in the Triangle of the Noghay Hordes, Crimean Khanate, and Ottoman Empire. Harvard Ukrainian Studies, 36(1-2), 105-122.', 'https://www.huri.harvard.edu/publications/harvard-ukrainian-studies', 'academic'),
+(338, 'Sinor, D. (ed.) (1990). The Cambridge History of Early Inner Asia. Cambridge University Press. ISBN 978-0521243049.', 'https://doi.org/10.1017/CHOL9780521243049', 'academic');
+
+-- Gobir (id=978) - 1000-1808 Hausa Bakwai
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(978, 'Hama, B. (1967). Histoire du Gobir et de Sokoto. Présence africaine.', 'https://www.presenceafricaine.com', 'academic'),
+(978, 'Hogben, S. J. & Kirk-Greene, A. H. M. (1966). The Emirates of Northern Nigeria: A Preliminary Survey of their Historical Traditions. Oxford University Press.', 'https://www.worldcat.org/title/emirates-of-northern-nigeria/oclc/250500', 'academic'),
+(978, 'Last, M. (2010). Ancient Labels and Categories: Exploring the Onomastics of Kano. In: A. Haour & B. Rossi (eds.), Being and Becoming Hausa: Interdisciplinary Studies, pp. 59-84. Brill. ISBN 978-9004185432.', 'https://brill.com/view/title/16770', 'academic'),
+(978, 'Adamu, M. (1984). The Hausa and their neighbours in the central Sudan. In: D. T. Niane (ed.), General History of Africa, Vol. IV. UNESCO/Heinemann. ISBN 978-9231017100.', 'https://unesdoc.unesco.org/ark:/48223/pf0000184282', 'academic'),
+(978, 'Lange, D. (1993). Die Hausa-Traditionen in ihrer Abhängigkeit von Kanem-Borno und Nubien. Anthropos, 88(1/3), 47-76.', 'https://www.jstor.org/stable/40463673', 'academic');
+
+-- Mafia Sultanate (id=1020) - 1100-1500 Swahili
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(1020, 'Chittick, H. N. (1961). Kisimani Mafia: Excavations at an Islamic Settlement on the East African Coast. Tanganyika Antiquities Department Occasional Paper 1.', 'https://www.worldcat.org/title/kisimani-mafia/oclc/8525461', 'academic'),
+(1020, 'Horton, M. & Middleton, J. (2000). The Swahili: The Social Landscape of a Mercantile Society. Blackwell. ISBN 978-0631189190.', 'https://www.wiley.com/en-us/The+Swahili-p-9780631189190', 'academic'),
+(1020, 'Pouwels, R. L. (1987). Horn and Crescent: Cultural Change and Traditional Islam on the East African Coast, 800-1900. Cambridge University Press. ISBN 978-0521523097.', 'https://www.cambridge.org/9780521523097', 'academic'),
+(1020, 'LaViolette, A. & Fleisher, J. B. (2018). The Swahili World. In: S. Wynne-Jones & A. LaViolette (eds.), The Swahili World, pp. 39-49. Routledge. ISBN 978-1138913462.', 'https://www.routledge.com/9781138913462', 'academic'),
+(1020, 'Wright, H. T. (1992). Early Islam, Oceanic Trade and Town Development on Nzwani: The Comorian Archipelago in the XIth-XVth Centuries AD. Azania, 27, 81-128. DOI:10.1080/00672709209511431.', 'https://doi.org/10.1080/00672709209511431', 'academic');
+
+-- Haudenosaunee proto-confederacy (id=941) - 800-1142
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(941, 'Trigger, B. G. (ed.) (1978). Handbook of North American Indians, Vol. 15: Northeast. Smithsonian Institution. ISBN 978-0160045752.', 'https://www.bookstore.gpo.gov', 'academic'),
+(941, 'Snow, D. R. (1994). The Iroquois. Blackwell. ISBN 978-1557863386.', 'https://www.wiley.com/en-us/The+Iroquois-p-9781557863386', 'academic'),
+(941, 'Engelbrecht, W. (2003). Iroquoia: The Development of a Native World. Syracuse University Press. ISBN 978-0815630142.', 'https://press.syr.edu/supressbooks/1162/iroquoia/', 'academic'),
+(941, 'Mann, B. A. & Fields, J. L. (1997). A Sign in the Sky: Dating the League of the Haudenosaunee. American Indian Culture and Research Journal, 21(2), 105-163.', 'https://escholarship.org/uc/uclajournals_aicrj/21/2', 'academic'),
+(941, 'Dennis, M. (1993). Cultivating a Landscape of Peace: Iroquois-European Encounters in Seventeenth-Century America. Cornell University Press. ISBN 978-0801480829.', 'https://www.cornellpress.cornell.edu/book/9780801480829/', 'academic');
+
+-- Duchy of Brabant (id=582) - 1183-1795
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(582, 'Arnade, P. (1996). Realms of Ritual: Burgundian Ceremony and Civic Life in Late Medieval Ghent. Cornell University Press. ISBN 978-0801432088.', 'https://www.cornellpress.cornell.edu/book/9780801432088/', 'academic'),
+(582, 'Blockmans, W. & Prevenier, W. (1999). The Promised Lands: The Low Countries Under Burgundian Rule, 1369-1530. University of Pennsylvania Press. ISBN 978-0812215724.', 'https://www.upenn.edu/pennpress/book/13902.html', 'academic'),
+(582, 'Stein, R. (2017). Magnanimous Dukes and Rising States: The Unification of the Burgundian Netherlands, 1380-1480. Oxford University Press. ISBN 978-0198757108.', 'https://global.oup.com/academic/product/magnanimous-dukes-and-rising-states-9780198757108', 'academic'),
+(582, 'Avonds, P. (1991). Brabant tijdens de regering van hertog Jan III (1312-1356). De grote politieke krisissen. Koninklijke Academie voor Wetenschappen, Letteren en Schone Kunsten. ISBN 978-9065690012.', 'https://www.kvab.be', 'academic'),
+(582, 'Israel, J. (1995). The Dutch Republic: Its Rise, Greatness and Fall 1477-1806. Oxford University Press. ISBN 978-0198730729.', 'https://global.oup.com/academic/product/the-dutch-republic-9780198730729', 'academic');
+
+-- Cirebon Sultanate (id=695) - 1527-1677 Javanese
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(695, 'Ricklefs, M. C. (2008). A History of Modern Indonesia Since c.1200 (4th ed.). Palgrave Macmillan. ISBN 978-0230546851.', 'https://link.springer.com/book/10.1057/9780230571631', 'academic'),
+(695, 'Guillot, C. (1990). The Sultanate of Banten. Gramedia Book Publishing Division. ISBN 978-9794031049.', 'https://www.gramedia.com', 'academic'),
+(695, 'Peacock, A. C. S. (2017). Islamisation: Comparative Perspectives from History. Edinburgh University Press. ISBN 978-1474417143.', 'https://edinburghuniversitypress.com/book-islamisation.html', 'academic'),
+(695, 'Muljana, S. (2005). Runtuhnya kerajaan Hindu-Jawa dan timbulnya negara-negara Islam di Nusantara. LKiS Pelangi Aksara. ISBN 978-9798451164.', 'https://lkis.co.id', 'academic'),
+(695, 'Noorduyn, J. & Teeuw, A. (2006). Three Old Sundanese Poems. KITLV Press / Brill. ISBN 978-9067183017.', 'https://brill.com/view/title/22904', 'academic');
+
+-- Confidence score upgrades
+UPDATE geo_entities SET confidence_score = 0.85 WHERE id IN (471, 473, 866, 643, 338, 978, 1020, 941, 582, 695);
+
+COMMIT;
+
+SELECT g.id, g.name_original, g.confidence_score, (SELECT count(*) FROM sources s WHERE s.entity_id=g.id) AS n_src FROM geo_entities g WHERE g.id IN (338, 471, 473, 582, 643, 695, 866, 941, 978, 1020) ORDER BY g.id;